% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{BasicAuth}
\alias{BasicAuth}
\title{BasicAuth Object}
\usage{
BasicAuth(password = NULL, user = NULL)
}
\arguments{
\item{password}{No description}

\item{user}{No description}
}
\value{
BasicAuth object
}
\description{
BasicAuth Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic Auth used as a credential.
}
