% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{CollectionOverride}
\alias{CollectionOverride}
\title{CollectionOverride Object}
\usage{
CollectionOverride(collection = NULL, options = NULL)
}
\arguments{
\item{collection}{The collection that identifies this resource within its service}

\item{options}{The options to apply to this resource-level override}
}
\value{
CollectionOverride object
}
\description{
CollectionOverride Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
CollectionOverride allows resource handling overrides for specific resources within a BaseType
}
