% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{CompositeType}
\alias{CompositeType}
\title{CompositeType Object}
\usage{
CompositeType(description = NULL, id = NULL, insertTime = NULL,
  labels = NULL, name = NULL, operation = NULL, selfLink = NULL,
  status = NULL, templateContents = NULL)
}
\arguments{
\item{description}{An optional textual description of the resource; provided by the client when the resource is created}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{insertTime}{[Output Only] Timestamp when the composite type was created, in RFC3339 text format}

\item{labels}{Map of labels; provided by the client when the resource is created or updated}

\item{name}{Name of the composite type}

\item{operation}{[Output Only] The Operation that most recently ran, or is currently running, on this composite type}

\item{selfLink}{[Output Only] Self link for the type provider}

\item{status}{No description}

\item{templateContents}{Files for the template type}
}
\value{
CompositeType object
}
\description{
CompositeType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Holds the composite type.
}
\seealso{
Other CompositeType functions: \code{\link{compositeTypes.insert}},
  \code{\link{compositeTypes.patch}},
  \code{\link{compositeTypes.update}}
}
