% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{CompositeTypeLabelEntry}
\alias{CompositeTypeLabelEntry}
\title{CompositeTypeLabelEntry Object}
\usage{
CompositeTypeLabelEntry(key = NULL, value = NULL)
}
\arguments{
\item{key}{No description}

\item{value}{No description}
}
\value{
CompositeTypeLabelEntry object
}
\description{
CompositeTypeLabelEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
