% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{CompositeTypesListResponse}
\alias{CompositeTypesListResponse}
\title{CompositeTypesListResponse Object}
\usage{
CompositeTypesListResponse(compositeTypes = NULL, nextPageToken = NULL)
}
\arguments{
\item{compositeTypes}{[Output Only] A list of resource composite types supported by Deployment Manager}

\item{nextPageToken}{A token used to continue a truncated list request}
}
\value{
CompositeTypesListResponse object
}
\description{
CompositeTypesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response that returns all Composite Types supported by Deployment Manager
}
