% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ConfigFile}
\alias{ConfigFile}
\title{ConfigFile Object}
\usage{
ConfigFile(content = NULL)
}
\arguments{
\item{content}{The contents of the file}
}
\value{
ConfigFile object
}
\description{
ConfigFile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
