% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ConfigurableService}
\alias{ConfigurableService}
\title{ConfigurableService Object}
\usage{
ConfigurableService(collectionOverrides = NULL, credential = NULL,
  descriptorUrl = NULL, options = NULL)
}
\arguments{
\item{collectionOverrides}{Allows resource handling overrides for specific collections}

\item{credential}{Credential used when interacting with this type}

\item{descriptorUrl}{Descriptor Url for the this type}

\item{options}{Options to apply when handling any resources in this service}
}
\value{
ConfigurableService object
}
\description{
ConfigurableService Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
BaseType that describes a service-backed Type.
}
