% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Credential}
\alias{Credential}
\title{Credential Object}
\usage{
Credential(basicAuth = NULL)
}
\arguments{
\item{basicAuth}{Basic Auth Credentials for this Type}
}
\value{
Credential object
}
\description{
Credential Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Credential used by ConfigurableResourceTypes.
}
