% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{DeploymentLabelEntry}
\alias{DeploymentLabelEntry}
\title{DeploymentLabelEntry Object}
\usage{
DeploymentLabelEntry(key = NULL, value = NULL)
}
\arguments{
\item{key}{No description}

\item{value}{No description}
}
\value{
DeploymentLabelEntry object
}
\description{
DeploymentLabelEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
