% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_functions.R
\docType{package}
\name{deploymentmanager_googleAuthR}
\alias{deploymentmanager_googleAuthR}
\alias{deploymentmanager_googleAuthR-package}
\title{Google Cloud Deployment Manager Alpha API
The Deployment Manager API allows users to declaratively configure, deploy and run complex solutions on the Google Cloud Platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:44:57
filename: /Users/mark/dev/R/autoGoogleAPI/googledeploymentmanageralpha.auto/R/deploymentmanager_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/ndev.cloudman
\item https://www.googleapis.com/auth/ndev.cloudman.readonly
}
}
