% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_functions.R
\name{deployments.delete}
\alias{deployments.delete}
\title{Deletes a deployment and all of the resources in the deployment.}
\usage{
deployments.delete(project, deployment, deletePolicy = NULL)
}
\arguments{
\item{project}{The project ID for this request}

\item{deployment}{The name of the deployment for this request}

\item{deletePolicy}{Sets the policy to use for deleting resources}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/}{Google Documentation}
}
