% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{DeploymentsListResponse}
\alias{DeploymentsListResponse}
\title{DeploymentsListResponse Object}
\usage{
DeploymentsListResponse(deployments = NULL, nextPageToken = NULL)
}
\arguments{
\item{deployments}{[Output Only] The deployments contained in this response}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}
}
\value{
DeploymentsListResponse object
}
\description{
DeploymentsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response containing a partial list of deployments and a page token used to build the next request if the request has been truncated.
}
