% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{DeploymentsStopRequest}
\alias{DeploymentsStopRequest}
\title{DeploymentsStopRequest Object}
\usage{
DeploymentsStopRequest(fingerprint = NULL)
}
\arguments{
\item{fingerprint}{Specifies a fingerprint for stop() requests}
}
\value{
DeploymentsStopRequest object
}
\description{
DeploymentsStopRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
\seealso{
Other DeploymentsStopRequest functions: \code{\link{deployments.stop}}
}
