% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{DeploymentUpdate}
\alias{DeploymentUpdate}
\title{DeploymentUpdate Object}
\usage{
DeploymentUpdate(description = NULL, labels = NULL, manifest = NULL)
}
\arguments{
\item{description}{[Output Only] An optional user-provided description of the deployment after the current update has been applied}

\item{labels}{[Output Only] Map of labels; provided by the client when the resource is created or updated}

\item{manifest}{[Output Only] URL of the manifest representing the update configuration of this deployment}
}
\value{
DeploymentUpdate object
}
\description{
DeploymentUpdate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
