% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{DeploymentUpdateLabelEntry}
\alias{DeploymentUpdateLabelEntry}
\title{DeploymentUpdateLabelEntry Object}
\usage{
DeploymentUpdateLabelEntry(key = NULL, value = NULL)
}
\arguments{
\item{key}{No description}

\item{value}{No description}
}
\value{
DeploymentUpdateLabelEntry object
}
\description{
DeploymentUpdateLabelEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
