% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{InputMapping}
\alias{InputMapping}
\title{InputMapping Object}
\usage{
InputMapping(fieldName = NULL, location = NULL, methodMatch = NULL,
  value = NULL)
}
\arguments{
\item{fieldName}{The name of the field that is going to be injected}

\item{location}{The location where this mapping applies}

\item{methodMatch}{Regex to evaluate on method to decide if input applies}

\item{value}{A jsonPath expression to select an element}
}
\value{
InputMapping object
}
\description{
InputMapping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
InputMapping creates a 'virtual' property that will be injected into the properties before sending the request to the underlying API.
}
