% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Manifest}
\alias{Manifest}
\title{Manifest Object}
\usage{
Manifest(config = NULL, expandedConfig = NULL, id = NULL,
  imports = NULL, insertTime = NULL, layout = NULL, name = NULL,
  selfLink = NULL)
}
\arguments{
\item{config}{[Output Only] The YAML configuration for this manifest}

\item{expandedConfig}{[Output Only] The fully-expanded configuration file, including any templates and references}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{imports}{[Output Only] The imported files for this manifest}

\item{insertTime}{[Output Only] Timestamp when the manifest was created, in RFC3339 text format}

\item{layout}{[Output Only] The YAML layout for this manifest}

\item{name}{[Output Only] The name of the manifest}

\item{selfLink}{[Output Only] Self link for the manifest}
}
\value{
Manifest object
}
\description{
Manifest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
