% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{OperationsListResponse}
\alias{OperationsListResponse}
\title{OperationsListResponse Object}
\usage{
OperationsListResponse(nextPageToken = NULL, operations = NULL)
}
\arguments{
\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}

\item{operations}{[Output Only] Operations contained in this list response}
}
\value{
OperationsListResponse object
}
\description{
OperationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response containing a partial list of operations and a page token used to build the next request if the request has been truncated.
}
