% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Options}
\alias{Options}
\title{Options Object}
\usage{
Options(inputMappings = NULL, nameProperty = NULL,
  validationOptions = NULL)
}
\arguments{
\item{inputMappings}{The mappings that apply for requests}

\item{nameProperty}{The json path to the field in the resource JSON body into which the resource name should be mapped}

\item{validationOptions}{Options for how to validate and process properties on a resource}
}
\value{
Options object
}
\description{
Options Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options allows customized resource handling by Deployment Manager.
}
