% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Resource}
\alias{Resource}
\title{Resource Object}
\usage{
Resource(Resource.warnings = NULL, Resource.warnings.data = NULL,
  accessControl = NULL, finalProperties = NULL, id = NULL,
  insertTime = NULL, manifest = NULL, name = NULL, properties = NULL,
  type = NULL, update = NULL, updateTime = NULL, url = NULL,
  warnings = NULL)
}
\arguments{
\item{Resource.warnings}{The \link{Resource.warnings} object or list of objects}

\item{Resource.warnings.data}{The \link{Resource.warnings.data} object or list of objects}

\item{accessControl}{The Access Control Policy set on this resource}

\item{finalProperties}{[Output Only] The evaluated properties of the resource with references expanded}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{insertTime}{[Output Only] Timestamp when the resource was created or acquired, in RFC3339 text format}

\item{manifest}{[Output Only] URL of the manifest representing the current configuration of this resource}

\item{name}{[Output Only] The name of the resource as it appears in the YAML config}

\item{properties}{[Output Only] The current properties of the resource before any references have been filled in}

\item{type}{[Output Only] The type of the resource, for example compute}

\item{update}{[Output Only] If Deployment Manager is currently updating or previewing an update to this resource, the updated configuration appears here}

\item{updateTime}{[Output Only] Timestamp when the resource was updated, in RFC3339 text format}

\item{url}{[Output Only] The URL of the actual resource}

\item{warnings}{[Output Only] If warning messages are generated during processing of this resource, this field will be populated}
}
\value{
Resource object
}
\description{
Resource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
\seealso{
Other Resource functions: \code{\link{Resource.warnings.data}},
  \code{\link{Resource.warnings}}
}
