% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ResourcesListResponse}
\alias{ResourcesListResponse}
\title{ResourcesListResponse Object}
\usage{
ResourcesListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{A token used to continue a truncated list request}

\item{resources}{Resources contained in this list response}
}
\value{
ResourcesListResponse object
}
\description{
ResourcesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response containing a partial list of resources and a page token used to build the next request if the request has been truncated.
}
