% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ResourceUpdate}
\alias{ResourceUpdate}
\title{ResourceUpdate Object}
\usage{
ResourceUpdate(ResourceUpdate.error = NULL,
  ResourceUpdate.error.errors = NULL, ResourceUpdate.warnings = NULL,
  ResourceUpdate.warnings.data = NULL, accessControl = NULL, error = NULL,
  finalProperties = NULL, intent = NULL, manifest = NULL,
  properties = NULL, state = NULL, warnings = NULL)
}
\arguments{
\item{ResourceUpdate.error}{The \link{ResourceUpdate.error} object or list of objects}

\item{ResourceUpdate.error.errors}{The \link{ResourceUpdate.error.errors} object or list of objects}

\item{ResourceUpdate.warnings}{The \link{ResourceUpdate.warnings} object or list of objects}

\item{ResourceUpdate.warnings.data}{The \link{ResourceUpdate.warnings.data} object or list of objects}

\item{accessControl}{The Access Control Policy to set on this resource after updating the resource itself}

\item{error}{[Output Only] If errors are generated during update of the resource, this field will be populated}

\item{finalProperties}{[Output Only] The expanded properties of the resource with reference values expanded}

\item{intent}{[Output Only] The intent of the resource: PREVIEW, UPDATE, or CANCEL}

\item{manifest}{[Output Only] URL of the manifest representing the update configuration of this resource}

\item{properties}{[Output Only] The set of updated properties for this resource, before references are expanded}

\item{state}{[Output Only] The state of the resource}

\item{warnings}{[Output Only] If warning messages are generated during processing of this resource, this field will be populated}
}
\value{
ResourceUpdate object
}
\description{
ResourceUpdate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
\seealso{
Other ResourceUpdate functions: \code{\link{ResourceUpdate.error.errors}},
  \code{\link{ResourceUpdate.error}},
  \code{\link{ResourceUpdate.warnings.data}},
  \code{\link{ResourceUpdate.warnings}}
}
