% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Rule}
\alias{Rule}
\title{Rule Object}
\usage{
Rule(action = NULL, conditions = NULL, description = NULL, ins = NULL,
  logConfigs = NULL, notIns = NULL, permissions = NULL)
}
\arguments{
\item{action}{Required}

\item{conditions}{Additional restrictions that must be met}

\item{description}{Human-readable description of the rule}

\item{ins}{If one or more 'in' clauses are specified, the rule matches if the PRINCIPAL/AUTHORITY_SELECTOR is in at least one of these entries}

\item{logConfigs}{The config returned to callers of tech}

\item{notIns}{If one or more 'not_in' clauses are specified, the rule matches if the PRINCIPAL/AUTHORITY_SELECTOR is in none of the entries}

\item{permissions}{A permission is a string of form '}
}
\value{
Rule object
}
\description{
Rule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rule to be applied in a Policy.
}
