% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{TemplateContents}
\alias{TemplateContents}
\title{TemplateContents Object}
\usage{
TemplateContents(imports = NULL, interpreter = NULL, schema = NULL,
  template = NULL)
}
\arguments{
\item{imports}{Import files referenced by the main template}

\item{interpreter}{Which interpreter (python or jinja) should be used during expansion}

\item{schema}{The contents of the template schema}

\item{template}{The contents of the main template file}
}
\value{
TemplateContents object
}
\description{
TemplateContents Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Files that make up the template contents of a template type.
}
