% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Type}
\alias{Type}
\title{Type Object}
\usage{
Type(configurableService = NULL, description = NULL, id = NULL,
  insertTime = NULL, labels = NULL, name = NULL, operation = NULL,
  selfLink = NULL)
}
\arguments{
\item{configurableService}{Base Type (configurable service) that backs this Type}

\item{description}{An optional textual description of the resource; provided by the client when the resource is created}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{insertTime}{[Output Only] Timestamp when the type was created, in RFC3339 text format}

\item{labels}{Map of labels; provided by the client when the resource is created or updated}

\item{name}{Name of the type}

\item{operation}{[Output Only] The Operation that most recently ran, or is currently running, on this type}

\item{selfLink}{[Output Only] Self link for the type}
}
\value{
Type object
}
\description{
Type Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A resource type supported by Deployment Manager.
}
\seealso{
Other Type functions: \code{\link{types.insert}},
  \code{\link{types.patch}}, \code{\link{types.update}}
}
