% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{TypeProvider}
\alias{TypeProvider}
\title{TypeProvider Object}
\usage{
TypeProvider(collectionOverrides = NULL, credential = NULL,
  description = NULL, descriptorUrl = NULL, id = NULL,
  insertTime = NULL, labels = NULL, name = NULL, operation = NULL,
  options = NULL, selfLink = NULL)
}
\arguments{
\item{collectionOverrides}{Allows resource handling overrides for specific collections}

\item{credential}{Credential used when interacting with this type}

\item{description}{An optional textual description of the resource; provided by the client when the resource is created}

\item{descriptorUrl}{Descriptor Url for the this type provider}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{insertTime}{[Output Only] Timestamp when the type provider was created, in RFC3339 text format}

\item{labels}{Map of labels; provided by the client when the resource is created or updated}

\item{name}{Name of the type provider}

\item{operation}{[Output Only] The Operation that most recently ran, or is currently running, on this type provider}

\item{options}{Options to apply when handling any resources in this service}

\item{selfLink}{[Output Only] Self link for the type provider}
}
\value{
TypeProvider object
}
\description{
TypeProvider Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A type provider that describes a service-backed Type.
}
\seealso{
Other TypeProvider functions: \code{\link{typeProviders.insert}},
  \code{\link{typeProviders.patch}},
  \code{\link{typeProviders.update}}
}
