% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{TypeProviderLabelEntry}
\alias{TypeProviderLabelEntry}
\title{TypeProviderLabelEntry Object}
\usage{
TypeProviderLabelEntry(key = NULL, value = NULL)
}
\arguments{
\item{key}{No description}

\item{value}{No description}
}
\value{
TypeProviderLabelEntry object
}
\description{
TypeProviderLabelEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
