% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_functions.R
\name{typeProviders.patch}
\alias{typeProviders.patch}
\title{Updates a type provider. This method supports patch semantics.}
\usage{
typeProviders.patch(TypeProvider, project, typeProvider)
}
\arguments{
\item{TypeProvider}{The \link{TypeProvider} object to pass to this method}

\item{project}{The project ID for this request}

\item{typeProvider}{The name of the type provider for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/}{Google Documentation}

Other TypeProvider functions: \code{\link{TypeProvider}},
  \code{\link{typeProviders.insert}},
  \code{\link{typeProviders.update}}
}
