% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{TypeProvidersListResponse}
\alias{TypeProvidersListResponse}
\title{TypeProvidersListResponse Object}
\usage{
TypeProvidersListResponse(nextPageToken = NULL, typeProviders = NULL)
}
\arguments{
\item{nextPageToken}{A token used to continue a truncated list request}

\item{typeProviders}{[Output Only] A list of resource type providers supported by Deployment Manager}
}
\value{
TypeProvidersListResponse object
}
\description{
TypeProvidersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response that returns all Type Providers supported by Deployment Manager
}
