% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_functions.R
\name{types.insert}
\alias{types.insert}
\title{Creates a type.}
\usage{
types.insert(Type, project)
}
\arguments{
\item{Type}{The \link{Type} object to pass to this method}

\item{project}{The project ID for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/}{Google Documentation}

Other Type functions: \code{\link{Type}},
  \code{\link{types.patch}}, \code{\link{types.update}}
}
