% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{TypesListResponse}
\alias{TypesListResponse}
\title{TypesListResponse Object}
\usage{
TypesListResponse(nextPageToken = NULL, types = NULL)
}
\arguments{
\item{nextPageToken}{A token used to continue a truncated list request}

\item{types}{[Output Only] A list of resource types supported by Deployment Manager}
}
\value{
TypesListResponse object
}
\description{
TypesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response that returns all Types supported by Deployment Manager
}
