% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ValidationOptions}
\alias{ValidationOptions}
\title{ValidationOptions Object}
\usage{
ValidationOptions(schemaValidation = NULL, undeclaredProperties = NULL)
}
\arguments{
\item{schemaValidation}{Customize how deployment manager will validate the resource against schema errors}

\item{undeclaredProperties}{Specify what to do with extra properties when executing a request}
}
\value{
ValidationOptions object
}
\description{
ValidationOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for how to validate and process properties on a resource.
}
